presentation screen now includes player colors
